pyproject.toml
[project]
name = "streams-charts-100-static"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]
